chore: core 역할에 따른 클래스명 지정 #328
