+++ backend/src/battles/core/battleMvp.ts
